Review comments reworked for yolo tutorials
