+++ python/nnvm/testing/__init__.py
@@ -8,3 +8,4 @@ from . import mlp
 from . import resnet
 from . import vgg
 from . import darknet
+from . import detection
